% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imported-and-re-exported.R
\name{seasonalview-exports}
\alias{as.yearmon}
\alias{as.yearqtr}
\alias{seasonalview-exports}
\title{imported and re exported functions from the zoo package}
\description{
The following functions are imported from zoo and re-exported from the 
seasonalview package.
}

