% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonalview-package.R
\docType{package}
\name{seasonalview-package}
\alias{seasonalview}
\alias{seasonalview-package}
\title{Graphical User Interface for Seasonal Adjustment}
\description{
A graphical user interface to the seasonal package and 
X-13ARIMA-SEATS, the U.S. Census Bureau's seasonal adjustment software. The
seasonalview package unifies the code base of 
\url{http://www.seasonal.website} with the GUI in the seasonal package.
}
\details{
Have a look at the README for usage examples: 
\url{https://github.com/christophsax/seasonalview}
}
\author{
Christoph Sax \email{christoph.sax@gmail.com}
}
\seealso{
\code{\link{view}} for the main function of seasonalview.
}
\keyword{package}

