% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone.R
\name{standalone}
\alias{standalone}
\title{Stand Alone Seasonal Web-Application}
\usage{
standalone()
}
\description{
Previews a stand alone web application for seasonal adjustment, whith the
same look as \url{www.seasonal.website}. Allows importing and exporting of 
data via the 'Up-/Download' button.
}
\examples{
\dontrun{
standalone()
}
}
\seealso{
\code{\link{view}} for an interactive GUI that imports and exports
  from the R console.
}

