% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewer.R
\name{view_story}
\alias{view_story}
\title{Local Display of Interactive Stories}
\usage{
view_story(file, quiet = TRUE)
}
\arguments{
\item{file}{character, path to the rmarkdown file containing an X-13 story.}

\item{quiet}{logical, should the output of shiny be suppressed}
}
\description{
Local Display of Interactive Stories
}
\examples{
\dontrun{
file <- system.file(package = "x13story", "stories", "x11.Rmd")
view(story = file)
}

}

