% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone.R
\name{standalone}
\alias{standalone}
\title{Stand Alone Seasonal Web-Application}
\usage{
standalone()
}
\description{
Previews a stand alone web application for seasonal adjustment, whith the
same look and feel as \url{http://www.seasonal.website}. Allows import and export of
data via an 'Up-/Download' button.
}
\examples{
\dontrun{
standalone()
}
}
\seealso{
\code{\link{view}} to interactively modify a \code{"seas"} object.
}
