% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view.R
\name{view}
\alias{view}
\title{Interactively Modify a Seasonal Adjustment Model}
\usage{
view(x = NULL, story = NULL, quiet = TRUE)
}
\arguments{
\item{x}{an object of class \code{"seas"}.}

\item{story}{character, path to an \code{".Rmd"} file.}

\item{quiet}{logical, if \code{TRUE} (default), error messages from calls in 
view are not shown in the console}
}
\value{
an object of class \code{"seas"}, the modified model.
}
\description{
Interactively modify a \code{"seas"} object. The goal of \code{view} is 
to summarize all relevant options, plots and statistics that should be 
usually considered.
}
\details{
Frequently used options can be modified using the drop down selectors in the
upper left window. Each change will result in a re-estimation of the seasonal
adjustment model. The R-call, the output and the summary are updated
accordingly.

Alternatively, the R-Call can be modified manually in the lower left window.
Press 'Run Call' to re-estimate the model and to adjust the option selectors,
the output, and the summary. With the 'to console' button, view is 
closed and the call is imported to R. The 'static' button substitutes 
automatic procedures by the automatically chosen 
spec-argument options, in the same way as \code{\link{static}}.

The views in the upper right window can be selected from the drop down menu.

The lower right panel shows the a summary, as described in the help page of
\code{\link{summary.seas}}. The 'Full X-13 output' button opens the complete 
output of X-13 in a separate tab or window.
}
\examples{
\dontrun{

m <- seas(AirPassengers)

view(m)

m2 <- view(m)  # save the model after closing the GUI

}
}

